# lib/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Rewards Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AuthStatusResponse {
  authenticated: boolean;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: string;
  name: string;
  slug: string;
  walletAddress: string;
  eligibleBalance: number;
  withdrawalFeeEth: number;
  feeWalletAddress: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateUserRequest {
  name: string;
  walletAddress: string;
  eligibleBalance: number;
  withdrawalFeeEth: number;
  feeWalletAddress: string;
}

export interface UpdateUserRequest {
  name?: string;
  walletAddress?: string;
  eligibleBalance?: number;
  withdrawalFeeEth?: number;
  feeWalletAddress?: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  commission: "commission",
  bonus: "bonus",
  withdrawal: "withdrawal",
  fee: "fee",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transaction {
  id: string;
  userId: string;
  amount: number;
  type: TransactionType;
  status: TransactionStatus;
  txHash?: string;
  note?: string;
  date: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTransactionRequestType =
  (typeof CreateTransactionRequestType)[keyof typeof CreateTransactionRequestType];

export const CreateTransactionRequestType = {
  commission: "commission",
  bonus: "bonus",
  withdrawal: "withdrawal",
  fee: "fee",
} as const;

export type CreateTransactionRequestStatus =
  (typeof CreateTransactionRequestStatus)[keyof typeof CreateTransactionRequestStatus];

export const CreateTransactionRequestStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface CreateTransactionRequest {
  amount: number;
  type: CreateTransactionRequestType;
  status: CreateTransactionRequestStatus;
  txHash?: string;
  note?: string;
  date: string;
}

export type UpdateTransactionRequestType =
  (typeof UpdateTransactionRequestType)[keyof typeof UpdateTransactionRequestType];

export const UpdateTransactionRequestType = {
  commission: "commission",
  bonus: "bonus",
  withdrawal: "withdrawal",
  fee: "fee",
} as const;

export type UpdateTransactionRequestStatus =
  (typeof UpdateTransactionRequestStatus)[keyof typeof UpdateTransactionRequestStatus];

export const UpdateTransactionRequestStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface UpdateTransactionRequest {
  amount?: number;
  type?: UpdateTransactionRequestType;
  status?: UpdateTransactionRequestStatus;
  txHash?: string;
  note?: string;
  date?: string;
}

export type WithdrawalRequestStatus =
  (typeof WithdrawalRequestStatus)[keyof typeof WithdrawalRequestStatus];

export const WithdrawalRequestStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface WithdrawalRequest {
  id: string;
  userId: string;
  userSlug: string;
  requestedAmount: number;
  feeAmount: number;
  status: WithdrawalRequestStatus;
  createdAt: string;
  updatedAt: string;
}

export type WithdrawalRequestOrNull = WithdrawalRequest | null;

export type WithdrawalRequestWithUserStatus =
  (typeof WithdrawalRequestWithUserStatus)[keyof typeof WithdrawalRequestWithUserStatus];

export const WithdrawalRequestWithUserStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface WithdrawalRequestWithUser {
  id: string;
  userId: string;
  userSlug: string;
  userName: string;
  requestedAmount: number;
  feeAmount: number;
  status: WithdrawalRequestWithUserStatus;
  createdAt: string;
  updatedAt: string;
}

export interface SubmitWithdrawalRequestBody {
  requestedAmount: number;
}