# lib/lib/api-zod/src/generated/types/createTransactionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Rewards Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionRequestStatus } from "./createTransactionRequestStatus";
import type { CreateTransactionRequestType } from "./createTransactionRequestType";

export interface CreateTransactionRequest {
  amount: number;
  type: CreateTransactionRequestType;
  status: CreateTransactionRequestStatus;
  txHash?: string;
  note?: string;
  date: string;
}